requirements.txt
dhash==1.3
pillow==7.2.0